backend/requirements.txt
fastapi==0.133.1
uvicorn==0.41.0
motor==3.7.1
pymongo==4.16.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-dotenv==1.2.1
PyGithub==2.8.1
httpx
python-multipart
resend==2.23.0
markdownify
requests==2.32.5
pydantic==2.12.5
pydantic-settings==2.13.1
email-validator==2.3.0
twilio==9.4.5
sendgrid==6.11.0
